requirements.txt
# FlagFoundry — Python Dependencies
# Training: 18× A6000 with DeepSpeed ZeRO-3
# Base model: Qwen2.5-7B-Coder-Instruct

# Core ML
torch>=2.3.0
transformers>=4.45.0
peft>=0.11.0
trl>=0.9.0
datasets>=2.20.0
accelerate>=0.30.0
deepspeed>=0.14.0
sentencepiece>=0.2.0
tokenizers>=0.19.0

# Inference serving
vllm>=0.4.0

# Anthropic API
anthropic>=0.28.0

# Data collection
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
lxml>=5.2.0

# CTF tooling
pwntools>=4.12.0
z3-solver>=4.13.0
pycryptodome>=3.20.0
scapy>=2.5.0

# Binary analysis
capstone>=5.0.0
unicorn>=2.0.1

# Web exploitation support
httpx>=0.27.0
websockets>=12.0

# Sandbox
docker>=7.0.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
loguru>=0.7.0
jsonlines>=4.0.0
datasketch>=1.6.5  # MinHash dedup

# Training monitoring
wandb>=0.16.6

# Serving
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Code analysis
tree-sitter>=0.21.0
pygments>=2.17.0

# Utils
python-dotenv>=1.0.1
pydantic>=2.6.0
rich>=13.7.0
typer>=0.12.0
tenacity>=8.2.3
PyYAML>=6.0.1
orjson>=3.10.0
pytest>=8.1.0
pytest-asyncio>=0.23.5

# LLM backends
openai>=1.20.0             # vLLM OpenAI-compatible API client
